requirements.txt
langchain
langchain-astradb
langchain-google-genai
google-generativeai
flask
pypdf
python-dotenv
flask

-e .
